feat: update component structures
