+++ src/app/(home)/_components/Pulse.tsx
@@ -3,15 +3,13 @@
 import { motion, useScroll } from 'framer-motion';
 import { type FC, useState } from 'react';
 
-const message = 'Technology, Design, and Community';
-
 export const Pulse: FC = () => {
   const { scrollY } = useScroll();
 
   const [fadeOut, setFadeOut] = useState(true);
 
   scrollY.updateAndNotify = (y: number) => {
-    if (y > innerHeight * 0.5 && y < innerHeight) {
+    if (y > innerHeight * 0.5 && y < innerHeight * 1.5) {
       setFadeOut(false);
     } else {
       setFadeOut(true);
@@ -20,7 +18,7 @@ export const Pulse: FC = () => {
 
   return (
     <motion.div
-      className="fixed h-64 w-full top-1/2 text-gray-300 bottom-0 font-bold leading-48 z-[-1]"
+      className="fixed h-32 w-full top-1/2 text-gray-300 bottom-0 font-bold leading-48 z-[-1]"
       initial={{ opacity: fadeOut ? 0 : 1 }}
       animate={{
         opacity: fadeOut ? 0 : 1,
@@ -28,7 +26,7 @@ export const Pulse: FC = () => {
       }}
     >
       <motion.div
-        className="w-full h-64 bg-gray-600"
+        className="w-full h-32 bg-gray-400"
         initial={{ backgroundPosition: '100% 50%' }}
         animate={{
           backgroundPosition: '0% 50%',
@@ -40,24 +38,20 @@ export const Pulse: FC = () => {
         style={{
           backgroundSize: '200% 100%',
           backgroundImage:
-            'linear-gradient(to right, #f6f3f4 45%, #193cb8 45%, #193cb8 50%, #f6f3f4 50%, #f6f3f4 100%)',
+            'linear-gradient(to right, #f6f3f4 45%, #0066cc 45%, #0066cc 50%, #f6f3f4 50%, #f6f3f4 100%)',
           clipPath: `polygon(
-            0% 50%,
-            20% 50%,
-            calc(20% + 10px) 98%,
-            calc(20% + 30px) 2%,
-            calc(20% + 38px) 70%,
-            calc(20% + 54px) 30%,
-            calc(20% + 62px) 50%,
-            100% 50%,
-            100% calc(50% + 2px),
-            calc(20% + 61px) calc(50% + 2px),
-            calc(20% + 54px) calc(30% + 6px),
-            calc(20% + 37px) calc(70% + 8px),
-            calc(20% + 30px) calc(2% + 12px),
-            calc(20% + 10px) calc(98% + 10px),
-            calc(20% - 1px) calc(50% + 2px),
-            0% calc(50% + 2px)
+            0 50%,
+            calc(10% + 2px) 50%,
+            calc(10% + 30px + 2px) 2px,
+            calc(10% + 90px - 2px) 2px,
+            calc(10% + 150px - 2px) 100%,
+            100% 100%,
+            100% calc(100% - 2px),
+            calc(10% + 150px) calc(100% - 2px),
+            calc(10% + 90px) 0,
+            calc(10% + 30px) 0,
+            10% calc(50% - 2px),
+            0 calc(50% - 2px)
           )`,
         }}
       />
